ocr_service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.2.0
opencv-python<=4.6.0.66
numpy==1.26.3
paddlepaddle==2.6.0
paddleocr==2.7.3
transformers==4.37.0
torch==2.5.1
sentencepiece==0.2.0
protobuf==3.20.2
packaging>=21.3
decorator>=5.0.0
pdf2image==1.16.3
